+++ src/app/app.component.ts
@@ -1,4 +1,6 @@
 import {AfterViewInit, Component, ElementRef, ViewChild} from '@angular/core';
+import {RouterOutlet} from '@angular/router';
+import {Title} from '@angular/platform-browser';
 
 @Component({
   selector: 'app-root',
@@ -8,7 +10,15 @@ import {AfterViewInit, Component, ElementRef, ViewChild} from '@angular/core';
 export class AppComponent implements AfterViewInit {
   @ViewChild('appTitle') title: ElementRef<HTMLHeadingElement>;
 
+  constructor(private titleService: Title) {
+  }
+
   ngAfterViewInit(): void {
     this.title.nativeElement.textContent = 'Shop';
+  }
+
+  onActivate($event: any, routerOutlet: RouterOutlet): void {
+    console.log('Activated Component', $event, routerOutlet);
+    this.titleService.setTitle(routerOutlet.activatedRouteData.title);
   }
 }