+++ src/app/app.module.ts
@@ -1,11 +1,15 @@
 import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
+import {Router} from '@angular/router';
 
 import {AppComponent} from './app.component';
 import {CartModule} from './cart/cart.module';
 import {FirstModule} from './first/first.module';
-import {ProductsModule} from './product-list/products.module';
 import {SharedModule} from './shared/shared.module';
+import {ProductsModule} from './products/products.module';
+import {AppRoutingModule} from './app-routing.module';
+import {OrdersModule} from './orders/orders.module';
+import {LayoutModule} from './layout/layout.module';
 
 @NgModule({
   declarations: [
@@ -14,13 +18,22 @@ import {SharedModule} from './shared/shared.module';
   imports: [
     BrowserModule,
 
+    LayoutModule,
+    SharedModule,
     CartModule,
     FirstModule,
     ProductsModule,
-    SharedModule
+    OrdersModule,
+    AppRoutingModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule {
+  constructor(router: Router) {
+    const replacer = (key: string, value: any): string =>
+      typeof value === 'function' ? value.name : value;
+
+    console.log('Routes: ', JSON.stringify(router.config, replacer, 2));
+  }
 }