+++ src/app/core/guards/auth.guard.ts
@@ -1,0 +1,46 @@
+import {Injectable} from '@angular/core';
+import {
+  CanActivate,
+  ActivatedRouteSnapshot,
+  RouterStateSnapshot,
+  UrlTree,
+  Router,
+  UrlSegment,
+  Route,
+  CanLoad
+} from '@angular/router';
+
+import {Observable} from 'rxjs';
+
+import {AuthService} from '../services';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthGuard implements CanActivate, CanLoad {
+  constructor(
+    private authService: AuthService,
+    private router: Router
+  ) {
+  }
+
+  canActivate(
+    route: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    return this.checkLogin(state.url);
+  }
+
+  canLoad(route: Route, segments: UrlSegment[]): Observable<boolean> | Promise<boolean> | boolean {
+    console.log('CanLoad Guard is called');
+    const url = `/${route.path}`;
+    return this.checkLogin(url) as boolean;
+  }
+
+  private checkLogin(url: string): boolean | UrlTree {
+    if (this.authService.isLoggedIn) {
+      return true;
+    }
+    this.authService.redirectUrl = url;
+    return this.router.parseUrl('/login');
+  }
+}