+++ src/app/core/services/index.ts
@@ -1,3 +1,4 @@
+export * from './auth.service';
 export * from './config-options.service';
 export * from './constant.service';
 export * from './generator.service';