+++ src/app/layout/components/index.ts
@@ -1,0 +1,2 @@
+export * from './login/login.component';
+export * from './path-not-found/path-not-found.component';