+++ src/app/layout/layout.module.ts
@@ -1,0 +1,13 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+
+import {LoginComponent, PathNotFoundComponent} from './components';
+
+@NgModule({
+  declarations: [PathNotFoundComponent, LoginComponent],
+  imports: [
+    CommonModule
+  ]
+})
+export class LayoutModule {
+}