+++ src/app/products/components/index.ts
@@ -1,0 +1,4 @@
+export * from './product/product.component';
+export * from './product-form/product-form.component';
+export * from './product-list/product-list.component';
+export * from './product-review/product-review.component';