+++ src/app/shared/directives/highlight.directive.spec.ts
@@ -1,9 +1,0 @@
-import { HighlightDirective } from './highlight.directive';
-import {ElementRef} from '@angular/core';
-
-describe('HighlightDirective', () => {
-  it('should create an instance', () => {
-    const directive = new HighlightDirective(new ElementRef<any>({}));
-    expect(directive).toBeTruthy();
-  });
-});