+++ src/app/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './models/app-enums.model';
+export * from './pipes/order-by.pipe';