+++ src/app/shared/shared.module.ts
@@ -2,8 +2,8 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 
-import {HighlightDirective} from './directives/highlight.directive';
-import {GoCrazyDirective} from './directives/go-crazy.directive';
+import {HighlightDirective} from './directives';
+import {GoCrazyDirective} from './directives';
 import {OrderByPipe} from './pipes/order-by.pipe';
 
 @NgModule({